FIX #14 - move gem files from `lib/` to `lib/geometry`
