+++ lib/geometry/algorithms/point_in_polygon.rb
