+++ lib/geometry/bounding_box.rb
