+++ lib/geometry/comparison_with_precision.rb
