+++ lib/geometry/line.rb
