+++ lib/geometry/point.rb
