+++ lib/geometry/polygon.rb
