+++ lib/geometry/segment.rb
