+++ lib/geometry/vector.rb
